Move to LLVM style RTTI
This is a very large change where we move from a variant based hierarchy to
a LLVM style RTTI hierarchy. This simplifies future integration with LLVM and use
of LLVM utilities like dyn_cast, etc.

It fixes a couple of outstanding bugs because it forces us to rethink casting between
types in the hierarchy.

Fixes Use LLVM RTTI style class hierarchy for AST Nodes #22
Fixes Templatize ToTopLevelNode and ToExprNode #14
Fixes Templatize downcast of ExprNode to ValueNode #30